pyproject.toml
[project]
name = "icocot-experiment"
version = "0.1.0"
description = "Invariant Counterfactual Checkpointing for Chain-of-Thought Reasoning"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.4.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
]